Change types and naming conventions

+++ src/WWS/Order/Dal/Orders/DbEntities/Order.cs
@@ -21,7 +21,7 @@ namespace Dal.Orders.DbEntities
         {
         }
 
-        public Order(long id, Guid orderCode, string customerId, string customerName, double totalPrice, ICollection<OrderItem> items, DateTime created, OrderState state)
+        public Order(long id, Guid orderCode, string customerId, string customerName, float totalPrice, ICollection<OrderItem> items, DateTime created, OrderState state)
         {
             _id = id;
             OrderCode = orderCode;
@@ -39,7 +39,7 @@ namespace Dal.Orders.DbEntities
         public Guid OrderCode { get; set; }
         public string CustomerId { get; set; }
         public string CustomerName { get; set; }
-        public double TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
         public ICollection<OrderItem> Items { get; set; } = new List<OrderItem>();
         public DateTime Created { get; set; }
         public OrderState State { get; set; }
